Moved models into the correct folder + removed the visualization handler and placed it into main profile handler
