+++ internal/database/models/visualization.go
@@ -1,4 +1,4 @@
-package profile
+package models
 
 import "time"
